+++ src/components/Calendar/CalendarRange/CalendarRange.tsx
@@ -35,9 +35,17 @@ export type CalendarRangeProps = React.HTMLAttributes<HTMLDivElement> & {
    */
   weekList?: string[];
   /**
+   * デフォルトで選択されているアクション
+   */
+  defaultClickAction?: string;
+  /**
    * カレンダーの左に表示するアクション
    */
   actions?: Action[];
+  /**
+   * アクションをクリックしたときの挙動
+   */
+  onClickAction?: (action: Action) => void;
   /**
    * 親コンポーネントで calendar を任意のタイミングで閉じたい場合に使用する
    */
@@ -69,9 +77,11 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
     {
       startDate,
       endDate,
-      monthFormat,
-      weekList,
+      monthFormat = "YYYY年M月",
+      weekList = ["日", "月", "火", "水", "木", "金", "土"],
+      defaultClickAction,
       actions,
+      onClickAction,
       onClose,
       isOutsideRange = () => false,
       onClickCloseButton,
@@ -125,7 +135,11 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
 
     return (
       <Card ref={ref} {...rest}>
-        <Actions actions={actions} />
+        <Actions
+          defaultClickAction={defaultClickAction}
+          actions={actions}
+          onClickAction={onClickAction}
+        />
         <Container>
           <div style={{ position: "relative", zIndex: 1 }}>
             <Slide unmountOnExit in={yearIsOpen} direction="up">