+++ src/components/Calendar/internal/Actions/styled.ts
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import ScrollArea from "../../../ScrollArea";
 
 export const Action = styled.button<{ clicked: boolean }>`
   cursor: pointer;
@@ -6,6 +7,12 @@ export const Action = styled.button<{ clicked: boolean }>`
   border: none;
   display: block;
   margin: 0;
+  max-width: 130px;
+  width: 100%;
+  text-align: left;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow: hidden;
   padding: ${({ theme }) => theme.spacing}px;
   color: ${({ clicked, theme }) =>
     clicked ? theme.palette.primary.main : theme.palette.black};
@@ -16,8 +23,6 @@ export const Action = styled.button<{ clicked: boolean }>`
   }
 `;
 
-export const ActionsContainer = styled.div`
-  width: 100%;
-  height: 100%;
+export const ActionsContainer = styled(ScrollArea)`
   padding: ${({ theme }) => `${theme.spacing}px ${theme.spacing * 2}px`};
 `;