+++ src/components/DateField/DateRangeField/DateRangeField.tsx
@@ -9,6 +9,7 @@ import {
   ClickStateType,
 } from "../../Calendar/CalendarRange/constants";
 import { useTheme } from "../../../themes";
+import { getInputWidth } from "../internal/utils";
 
 export type DateRangeFieldProps = {
   /**
@@ -52,7 +53,7 @@ const DateRangeField = forwardRef<HTMLInputElement, DateRangeFieldProps>(
     propRef,
   ) {
     const theme = useTheme();
-    const width = useMemo(() => format.length * 12, [format]);
+    const width = useMemo(() => `${getInputWidth(format) * 10}px`, [format]);
 
     const handleChange = (t: ClickStateType) => (date: Dayjs) => {
       const { startDate, endDate } = rest.date;